+++ library/hexaminate/lib/extension/bool.dart
@@ -3,9 +3,3 @@ part of hexaminate;
 extension BoolEncodeDecodeExtension on bool {
 
 }
-
-extension BoolToBooleanExtension on bool {
-  bool get toBoolean {
-    return this;
-  }
-}