+++ library/hexaminate/lib/extension/dynamic.dart
@@ -1,1 +1,25 @@
 part of hexaminate;
+
+
+extension DynamicUtilsTypeExntension on dynamic {
+  String get type {
+    return typeof(this);
+  }
+
+  bool get toBoolean {
+    if (this == null) {
+      return false;
+    }
+    return getBoolean(this);
+  }
+}
+
+extension NullableTypeExntensiont on Null {
+  String get type {
+    return typeof(this);
+  }
+
+  bool get toBoolean {
+    return false;
+  }
+}