+++ library/hexaminate/lib/extension/map.dart
@@ -1,7 +1,5 @@
 part of hexaminate;
 
-extension EncodeDecode<A, V> on Map<A, V> {}
-
 extension Loop<K, V> on Map<K, V> {
   void loop(void Function(K key, V value, int index) callback) {
     var i = 0;
@@ -22,11 +20,5 @@ extension IsType<K, V> on Map<K, V> {
         .replaceAll(RegExp(r"<.*"), "")
         .replaceAll(RegExp(r"_internallinkedhashmap"), "object")
         .replaceAll(RegExp(r"_"), "");
-  }
-}
-
-extension ToBoolean<K, V> on Map<K, V> {
-  bool get toBoolean {
-    return isNotEmpty;
   }
 }